# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Management System API",
        "description": "API for managing hospital operations with role-based access control",
        "contact": {
            "email": "admin@hospital.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/admin/medicines": {
            "get": {
                "tags": [
                    "⚙️ Admin Management"
                ],
                "summary": "Get all medicines",
                "description": "Retrieves all medicines in the hospital inventory (admin only)",
                "operationId": "getAllMedicines",
                "responses": {
                    "200": {
                        "description": "Medicines retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Medicines retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "medicines": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/MedicineResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "⚙️ Admin Management"
                ],
                "summary": "Create new medicine",
                "description": "Creates a new medicine in the inventory (admin only)",
                "operationId": "createMedicine",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "medicine_name",
                                    "medicine_type",
                                    "dosage",
                                    "unit",
                                    "stock"
                                ],
                                "properties": {
                                    "medicine_name": {
                                        "type": "string",
                                        "example": "Amlodipine"
                                    },
                                    "medicine_type": {
                                        "type": "string",
                                        "example": "tablet"
                                    },
                                    "dosage": {
                                        "type": "string",
                                        "example": "5mg"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "tablet"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Calcium channel blocker for hypertension"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Medicine created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Medicine created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "medicine": {
                                                    "$ref": "#/components/schemas/MedicineResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/doctor/medicines": {
            "get": {
                "tags": [
                    "👩‍⚕️ Doctor Services"
                ],
                "summary": "Get available medicines",
                "description": "Retrieves all medicines with stock > 0 for doctors to prescribe",
                "operationId": "getDoctorAvailableMedicines",
                "responses": {
                    "200": {
                        "description": "Available medicines retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Available medicines retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "medicines": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/MedicineResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/admin/medicines/{id}/stock": {
            "put": {
                "tags": [
                    "⚙️ Admin Management"
                ],
                "summary": "Update medicine stock",
                "description": "Updates medicine stock quantity (admin only)",
                "operationId": "updateMedicineStock",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Medicine ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "stock"
                                ],
                                "properties": {
                                    "stock": {
                                        "description": "New stock quantity",
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "reason": {
                                        "description": "Reason for stock update",
                                        "type": "string",
                                        "example": "Stock replenishment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Stock updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "medicine": {
                                                    "$ref": "#/components/schemas/MedicineResource"
                                                },
                                                "previous_stock": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "new_stock": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MedicineResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "medicine_name": {
                        "type": "string",
                        "example": "Amlodipine"
                    },
                    "medicine_type": {
                        "type": "string",
                        "example": "tablet"
                    },
                    "dosage": {
                        "type": "string",
                        "example": "5mg"
                    },
                    "unit": {
                        "type": "string",
                        "example": "tablet"
                    },
                    "stock": {
                        "type": "integer",
                        "example": 100
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}